third_party/blink/web_tests/virtual/text-antialias/find-kana-expected.txt

Tests find for strings with kana letters in them.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Exact matches first as a baseline


PASS canFind(decomposedHalfwidthKatakanaLetterBa, decomposedHalfwidthKatakanaLetterBa) is true
PASS canFind(decomposedHalfwidthKatakanaLetterPa, decomposedHalfwidthKatakanaLetterPa) is true
PASS canFind(decomposedHiraganaLetterBa, decomposedHiraganaLetterBa) is true
PASS canFind(decomposedHiraganaLetterGa, decomposedHiraganaLetterGa) is true
PASS canFind(decomposedHiraganaLetterPa, decomposedHiraganaLetterPa) is true
PASS canFind(decomposedKatakanaLetterGa, decomposedKatakanaLetterGa) is true
PASS canFind(decomposedLatinCapitalLetterAWithGrave, decomposedLatinCapitalLetterAWithGrave) is true
PASS canFind(halfwidthKatakanaLetterA, halfwidthKatakanaLetterA) is true
PASS canFind(halfwidthKatakanaLetterHa, halfwidthKatakanaLetterHa) is true
PASS canFind(halfwidthKatakanaLetterKa, halfwidthKatakanaLetterKa) is true
PASS canFind(halfwidthKatakanaLetterSmallA, halfwidthKatakanaLetterSmallA) is true
PASS canFind(hiraganaLetterA, hiraganaLetterA) is true
PASS canFind(hiraganaLetterBa, hiraganaLetterBa) is true
PASS canFind(hiraganaLetterGa, hiraganaLetterGa) is true
PASS canFind(hiraganaLetterHa, hiraganaLetterHa) is true
PASS canFind(hiraganaLetterKa, hiraganaLetterKa) is true
PASS canFind(hiraganaLetterPa, hiraganaLetterPa) is true
PASS canFind(hiraganaLetterSmallA, hiraganaLetterSmallA) is true
PASS canFind(katakanaLetterA, katakanaLetterA) is true
PASS canFind(katakanaLetterGa, katakanaLetterGa) is true
PASS canFind(katakanaLetterKa, katakanaLetterKa) is true
PASS canFind(katakanaLetterSmallA, katakanaLetterSmallA) is true
PASS canFind(latinCapitalLetterAWithGrave, latinCapitalLetterAWithGrave) is true


Hiragana, katakana, and half width katakana: Must be treated as equal


PASS canFind(decomposedHiraganaLetterGa, decomposedKatakanaLetterGa) is true
PASS canFind(decomposedKatakanaLetterGa, decomposedHiraganaLetterGa) is true
PASS canFind(hiraganaLetterA, halfwidthKatakanaLetterA) is true
PASS canFind(hiraganaLetterA, katakanaLetterA) is true
PASS canFind(katakanaLetterSmallA, hiraganaLetterSmallA) is true


Composed and decomposed forms: Must be treated as equal


PASS canFind(decomposedHiraganaLetterBa, hiraganaLetterBa) is true
PASS canFind(decomposedHiraganaLetterGa, decomposedKatakanaLetterGa) is true
PASS canFind(decomposedHiraganaLetterGa, hiraganaLetterGa) is true
PASS canFind(decomposedHiraganaLetterGa, katakanaLetterGa) is true
PASS canFind(decomposedHiraganaLetterPa, hiraganaLetterPa) is true
PASS canFind(decomposedKatakanaLetterGa, decomposedHiraganaLetterGa) is true
PASS canFind(decomposedLatinCapitalLetterAWithGrave, latinCapitalLetterAWithGrave) is true
PASS canFind(hiraganaLetterBa, decomposedHiraganaLetterBa) is true
PASS canFind(hiraganaLetterGa, decomposedHiraganaLetterGa) is true
PASS canFind(hiraganaLetterPa, decomposedHiraganaLetterPa) is true
PASS canFind(katakanaLetterGa, decomposedHiraganaLetterGa) is true
PASS canFind(latinCapitalLetterAWithGrave, decomposedLatinCapitalLetterAWithGrave) is true


Small and non-small kana letters: Must *not* be treated as equal


PASS canFind(halfwidthKatakanaLetterA, hiraganaLetterSmallA) is false
PASS canFind(halfwidthKatakanaLetterSmallA, halfwidthKatakanaLetterA) is false
PASS canFind(hiraganaLetterA, hiraganaLetterSmallA) is false
PASS canFind(hiraganaLetterSmallA, katakanaLetterA) is false
PASS canFind(katakanaLetterA, halfwidthKatakanaLetterSmallA) is false
PASS canFind(katakanaLetterSmallA, katakanaLetterA) is false


Kana letters where the only difference is in voiced sound marks: Must *not* be treated as equal


PASS canFind(decomposedHalfwidthKatakanaLetterBa, halfwidthKatakanaLetterHa) is false
PASS canFind(decomposedHalfwidthKatakanaLetterPa, halfwidthKatakanaLetterHa) is false
PASS canFind(decomposedHiraganaLetterBa, hiraganaLetterHa) is false
PASS canFind(decomposedHiraganaLetterBa, hiraganaLetterPa) is false
PASS canFind(decomposedHiraganaLetterGa, halfwidthKatakanaLetterKa) is false
PASS canFind(decomposedHiraganaLetterGa, hiraganaLetterKa) is false
PASS canFind(decomposedHiraganaLetterPa, hiraganaLetterBa) is false
PASS canFind(decomposedHiraganaLetterPa, hiraganaLetterHa) is false
PASS canFind(decomposedKatakanaLetterGa, halfwidthKatakanaLetterKa) is false
PASS canFind(decomposedKatakanaLetterGa, hiraganaLetterKa) is false
PASS canFind(halfwidthKatakanaLetterHa, decomposedHalfwidthKatakanaLetterBa) is false
PASS canFind(halfwidthKatakanaLetterHa, decomposedHalfwidthKatakanaLetterPa) is false
PASS canFind(halfwidthKatakanaLetterKa, decomposedHiraganaLetterGa) is false
PASS canFind(halfwidthKatakanaLetterKa, decomposedKatakanaLetterGa) is false
PASS canFind(hiraganaLetterBa, decomposedHiraganaLetterPa) is false
PASS canFind(hiraganaLetterBa, hiraganaLetterHa) is false
PASS canFind(hiraganaLetterBa, hiraganaLetterPa) is false
PASS canFind(hiraganaLetterGa, hiraganaLetterKa) is false
PASS canFind(hiraganaLetterHa, decomposedHiraganaLetterBa) is false
PASS canFind(hiraganaLetterHa, decomposedHiraganaLetterPa) is false
PASS canFind(hiraganaLetterHa, hiraganaLetterBa) is false
PASS canFind(hiraganaLetterHa, hiraganaLetterPa) is false
PASS canFind(hiraganaLetterKa, decomposedHiraganaLetterGa) is false
PASS canFind(hiraganaLetterKa, decomposedKatakanaLetterGa) is false
PASS canFind(hiraganaLetterKa, hiraganaLetterGa) is false
PASS canFind(hiraganaLetterPa, decomposedHiraganaLetterBa) is false
PASS canFind(hiraganaLetterPa, hiraganaLetterBa) is false
PASS canFind(hiraganaLetterPa, hiraganaLetterHa) is false


Composed/decomposed form differences before kana characters must have no effect


PASS canFind(decomposedLatinCapitalLetterAWithGrave + halfwidthKatakanaLetterA, latinCapitalLetterAWithGrave + hiraganaLetterSmallA) is false
PASS canFind(decomposedLatinCapitalLetterAWithGrave + halfwidthKatakanaLetterSmallA, latinCapitalLetterAWithGrave + halfwidthKatakanaLetterA) is false
PASS canFind(decomposedLatinCapitalLetterAWithGrave + hiraganaLetterA, latinCapitalLetterAWithGrave + hiraganaLetterSmallA) is false
PASS canFind(decomposedLatinCapitalLetterAWithGrave + hiraganaLetterGa, latinCapitalLetterAWithGrave + hiraganaLetterGa) is true
PASS canFind(decomposedLatinCapitalLetterAWithGrave + hiraganaLetterGa, latinCapitalLetterAWithGrave + hiraganaLetterKa) is false
PASS canFind(decomposedLatinCapitalLetterAWithGrave + hiraganaLetterKa, latinCapitalLetterAWithGrave + hiraganaLetterGa) is false
PASS canFind(decomposedLatinCapitalLetterAWithGrave + hiraganaLetterSmallA, latinCapitalLetterAWithGrave + katakanaLetterA) is false
PASS canFind(decomposedLatinCapitalLetterAWithGrave + katakanaLetterA, latinCapitalLetterAWithGrave + halfwidthKatakanaLetterSmallA) is false
PASS canFind(decomposedLatinCapitalLetterAWithGrave + katakanaLetterSmallA, latinCapitalLetterAWithGrave + katakanaLetterA) is false
PASS canFind(latinCapitalLetterAWithGrave + halfwidthKatakanaLetterA, decomposedLatinCapitalLetterAWithGrave + hiraganaLetterSmallA) is false
PASS canFind(latinCapitalLetterAWithGrave + halfwidthKatakanaLetterSmallA, decomposedLatinCapitalLetterAWithGrave + halfwidthKatakanaLetterA) is false
PASS canFind(latinCapitalLetterAWithGrave + hiraganaLetterA, decomposedLatinCapitalLetterAWithGrave + hiraganaLetterSmallA) is false
PASS canFind(latinCapitalLetterAWithGrave + hiraganaLetterGa, decomposedLatinCapitalLetterAWithGrave + hiraganaLetterGa) is true
PASS canFind(latinCapitalLetterAWithGrave + hiraganaLetterGa, decomposedLatinCapitalLetterAWithGrave + hiraganaLetterKa) is false
PASS canFind(latinCapitalLetterAWithGrave + hiraganaLetterKa, decomposedLatinCapitalLetterAWithGrave + hiraganaLetterGa) is false
PASS canFind(latinCapitalLetterAWithGrave + hiraganaLetterSmallA, decomposedLatinCapitalLetterAWithGrave + katakanaLetterA) is false
PASS canFind(latinCapitalLetterAWithGrave + katakanaLetterA, decomposedLatinCapitalLetterAWithGrave + halfwidthKatakanaLetterSmallA) is false
PASS canFind(latinCapitalLetterAWithGrave + katakanaLetterSmallA, decomposedLatinCapitalLetterAWithGrave + katakanaLetterA) is false


PASS successfullyParsed is true

TEST COMPLETE
